refactor: 컴포넌트 파일 절대경로 수정
