+++ frontend/src/components/index.ts
@@ -1,41 +1,42 @@
-export { default as Button } from './Button/Button';
-export { default as Dropdown } from './Dropdown/Dropdown';
-export { default as Flex } from './Flex/Flex';
-export { default as Header } from './Header/Header';
-export { default as Heading } from './Heading/Heading';
-export { default as Input } from './Input/Input';
-export { default as Layout } from './Layout/Layout';
-export { default as LikeButton } from './LikeButton/LikeButton';
-export { default as LikeCounter } from './LikeCounter/LikeCounter';
-export { default as Modal } from './Modal/Modal';
-export { default as NonmemberAlerter } from './NonmemberAlerter/NonmemberAlerter';
-export { default as PagingButtons } from './PagingButtons/PagingButtons';
-export { default as SelectList } from './SelectList/SelectList';
-export { default as SourceCode } from './SourceCode/SourceCode';
-export { default as SourceCodeEditor } from './SourceCodeEditor/SourceCodeEditor';
-export { default as SourceCodeViewer } from './SourceCodeViewer/SourceCodeViewer';
-export { default as TagButton } from './TagButton/TagButton';
-export { default as TagInput } from './TagInput/TagInput';
-export { default as TemplateCard } from './TemplateCard/TemplateCard';
-export { default as Text } from './Text/Text';
-export { default as Toast } from './Toast/Toast';
-export { default as Guide } from './Guide/Guide';
-export { default as Footer } from './Footer/Footer';
-export { default as CategoryDropdown } from './CategoryDropdown/CategoryDropdown';
-export { default as CategoryGuide } from './CategoryGuide/CategoryGuide';
-export { default as NewCategoryInput } from './NewCategoryInput/NewCategoryInput';
-export { default as NoResults } from './NoResults/NoResults';
-export { default as Textarea } from './Textarea/Textarea';
-export { default as ContactUs } from './ContactUs/ContactUs';
-export { default as Toggle } from './Toggle/Toggle';
-export { default as ScrollTopButton } from './ScrollTopButton/ScrollTopButton';
+export { default as Button } from '@/components/Button/Button';
+export { default as Dropdown } from '@/components/Dropdown/Dropdown';
+export { default as Flex } from '@/components/Flex/Flex';
+export { default as Header } from '@/components/Header/Header';
+export { default as Heading } from '@/components/Heading/Heading';
+export { default as Input } from '@/components/Input/Input';
+export { default as Layout } from '@/components/Layout/Layout';
+export { default as LikeButton } from '@/components/LikeButton/LikeButton';
+export { default as LikeCounter } from '@/components/LikeCounter/LikeCounter';
+export { default as Modal } from '@/components/Modal/Modal';
+export { default as NonmemberAlerter } from '@/components/NonmemberAlerter/NonmemberAlerter';
+export { default as PagingButtons } from '@/components/PagingButtons/PagingButtons';
+export { default as SelectList } from '@/components/SelectList/SelectList';
+export { default as SourceCode } from '@/components/SourceCode/SourceCode';
+export { default as SourceCodeEditor } from '@/components/SourceCodeEditor/SourceCodeEditor';
+export { default as SourceCodeViewer } from '@/components/SourceCodeViewer/SourceCodeViewer';
+export { default as TagButton } from '@/components/TagButton/TagButton';
+export { default as TagInput } from '@/components/TagInput/TagInput';
+export { default as TemplateCard } from '@/components/TemplateCard/TemplateCard';
+export { default as Text } from '@/components/Text/Text';
+export { default as Toast } from '@/components/Toast/Toast';
+export { default as Guide } from '@/components/Guide/Guide';
+export { default as Footer } from '@/components/Footer/Footer';
+export { default as CategoryDropdown } from '@/components/CategoryDropdown/CategoryDropdown';
+export { default as CategoryGuide } from '@/components/CategoryGuide/CategoryGuide';
+export { default as NewCategoryInput } from '@/components/NewCategoryInput/NewCategoryInput';
+export { default as NoResults } from '@/components/NoResults/NoResults';
+export { default as Textarea } from '@/components/Textarea/Textarea';
+export { default as ContactUs } from '@/components/ContactUs/ContactUs';
+export { default as Toggle } from '@/components/Toggle/Toggle';
+export { default as ScrollTopButton } from '@/components/ScrollTopButton/ScrollTopButton';
+export { default as AuthorInfo } from '@/components/AuthorInfo/AuthorInfo';
 
 // Skeleton UI
-export { default as LoadingBall } from './LoadingBall/LoadingBall';
-export { default as LoadingFallback } from './LoadingFallback/LoadingFallback';
+export { default as LoadingBall } from '@/components/LoadingBall/LoadingBall';
+export { default as LoadingFallback } from '@/components/LoadingFallback/LoadingFallback';
 
 // Error UI
-export { default as TemporaryError } from './Error/TemporaryError';
+export { default as TemporaryError } from '@/components/Error/TemporaryError';
 
 // ScreenReader
-export { default as ScreenReaderOnly } from './ScreenReaderOnly/ScreenReaderOnly';
+export { default as ScreenReaderOnly } from '@/components/ScreenReaderOnly/ScreenReaderOnly';
